refactor: 자판기 클래스 쪼개기
