+++ src/js/service/VendingMachine.js
@@ -1,18 +1,13 @@
-import { ALERT_MESSAGE, VENDING_MACHINE_CONSTANT } from './constant.js';
+import { ALERT_MESSAGE } from './constant.js';
 import { removeSpace } from '../util/string.js';
-import { isAmountValid, isInsertedCoinsValid, isNameValid, isPriceValid } from './validator.js';
+import { isAmountValid, isNameValid, isPriceValid } from './validator.js';
 import ValidationError from './ValidationError.js';
-
-/**
- * @typedef {Object} UnitCountInfo
- * @property {number} amount
- * @property {Object.<string, number>} unitInfo
- */
+import UnitCountMachine from './UnitCountMachine.js';
 
 export class VendingMachine {
   #products;
-  /**@type {UnitCountInfo} */
-  #unitCountInfo;
+  /**@type {UnitCountMachine} */
+  #unitCountMachine;
 
   constructor() {
     this.reset();
@@ -20,37 +15,19 @@ export class VendingMachine {
 
   reset() {
     this.#products = [];
-    this.#unitCountInfo = {
-      amount: 0,
-      unitInfo: VENDING_MACHINE_CONSTANT.CHANGES.UNITS.reduce(
-        (result, unit) => ({
-          ...result,
-          [unit]: 0,
-        }),
-        {}
-      ),
-    };
-    console.log(this.#unitCountInfo);
+    this.#unitCountMachine = new UnitCountMachine();
   }
 
   /**
    *
    * @returns {VendingMachineItem[]}
    */
-  getProducts() {
-    return [...this.#products];
+  get products() {
+    return this.#products;
   }
 
-  /**
-   *
-   * @returns {UnitCountInfo}
-   */
-  get unitCountInfo() {
-    return this.#unitCountInfo;
-  }
-
-  getUnits() {
-    return Object.keys(this.#unitCountInfo.unitInfo).sort((a, b) => Number(b) - Number(a));
+  get unitCountMachine() {
+    return this.#unitCountMachine;
   }
 
   /**
@@ -87,53 +64,6 @@ export class VendingMachine {
     if (!isAmountValid(amount)) {
       throw new ValidationError(ALERT_MESSAGE.VALIDATION.PRODUCT.AMOUNT);
     }
-  }
-
-  get #units() {
-    return Object.keys(this.#unitCountInfo.unitInfo).sort((a, b) => Number(b) - Number(a));
-  }
-
-  /**
-   *
-   * @param {number} amount
-   * @returns {UnitCountInfo}
-   */
-  #getUnitCountInfo(chargeAmount) {
-    const { unitInfo } = this.#units.reduce(
-      ({ amount, unitInfo }, unit) => {
-        const count = Math.floor(amount / unit);
-        return {
-          amount: amount - count * unit,
-          unitInfo: { ...unitInfo, [unit]: count },
-        };
-      },
-      { amount: chargeAmount, unitInfo: {} }
-    );
-    return { amount: chargeAmount, unitInfo };
-  }
-
-  #validateCoins(amount) {
-    if (!isInsertedCoinsValid(amount)) {
-      throw new ValidationError(ALERT_MESSAGE.VALIDATION.CHARGE_AMOUNT);
-    }
-  }
-
-  /**
-   *
-   * @param {number} amount
-   */
-  insertCoins(amount) {
-    this.#validateCoins(amount);
-    const insertedUnitCountInfo = this.#getUnitCountInfo(amount);
-    this.#unitCountInfo = {
-      amount: this.#unitCountInfo.amount + insertedUnitCountInfo.amount,
-      unitInfo: Object.keys(insertedUnitCountInfo.unitInfo).reduce(
-        (result, unit) => {
-          return { ...result, [unit]: insertedUnitCountInfo.unitInfo[unit] + this.#unitCountInfo.unitInfo[unit] };
-        },
-        { ...this.#unitCountInfo.unitInfo }
-      ),
-    };
   }
 }
